feature: rename mapper
