+++ app/src/main/java/com/hacka/presenteperfeito/feature/signup/presentation/SignUpViewModel.kt
@@ -6,7 +6,7 @@ import com.hacka.presenteperfeito.R
 import com.hacka.presenteperfeito.core.common.BaseViewModel
 import com.hacka.presenteperfeito.core.common.validator.FormValidator
 import com.hacka.presenteperfeito.feature.signup.domain.usecase.SignUpUseCase
-import com.hacka.presenteperfeito.feature.signup.presentation.model.CreateUser
+import com.hacka.presenteperfeito.feature.signup.domain.model.CreateUser
 import com.hacka.presenteperfeito.feature.signup.presentation.uiState.SignUpFormState
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.launch
